Put progress log from xtrabackup in log file instead of throwing it to /dev/null
When Popen is called with stderr=None, the child will get a
/dev/null connected to its stderr, and communicate() will not be
able to return any output from it.

Instead this patch introduces a new directive, log=filename,
which by default is enabled and set to "plugin-percona.log".
This file is created in bareos-fd's working dir (typically
/var/lib/bareos) unless the filename is absolute.  The special
value "false" will disable logging ("/dev/null" will have the
same effect).

To avoid disks filling up, the file is opened for writing, not
appending, on each run.

This patch also fixes the usage of MySQLdb so that it reads
.my.cnf correctly even when mycnf parameter is left out.