This repository has been merged into https://github.com/bareos/bareos
The Bareos project has decided to merge its sub projects into the main repository.

  * https://github.com/bareos/bareos-contrib (master) =>
    https://github.com/bareos/bareos/tree/master/contrib/

This makes it easier to keep the code in sync for the Bareos releases.

We only merged the master branch. So this and all future branches from
bareos-21 onwards can be found there.
The old branch bareos-19.2 will be kept in the old repositories.